Exploratory data analysis on call center
Problem Background
CallIn PLC is a BPO solutions company based in Poland that handles incoming calls for various banks. They have dedicated teams for each of their clients. This case study is about the team that handles incoming calls for Royal Bank of Scotland (RBS) that is one of the banks that employs service solutions of CallIn PLC.