Problem statement
You have been given data of medicine sales of the year 2020 and 2021 for the country. The data includes the sale of medicines through app and website which are segmented as “Retail” channels. The company now wants to expand into “Institutional” segment that could lead to large scale procurement of medicines from them. The institutional segment could also utilize online as well as offline sales channels.
As per information received, multiple number of bulk ordering agents are already utilizing the services of Medplus. The business managers at the company have taken a ballpark approximate figure of Rs 2,000/ transaction to qualify for being a bulk order. Hence, the given data of 65,969 rows and 10 columns includes only the customer details that has each transaction value of more then 2,000.