Crazy Monkey Omnichannel platform EDA
The company wants to expand across India through an online web market and various offline stores. Still, before taking this step, they wanted to be double sure regarding the demand and the sector they should target. The company has a limited cash reserve and they want to utilize it in the best possible way. 

The company has hired you as an External Data Analyst and wanted you to help them with the following questions:
1.	The top 3 cities that the company should target. Suggest where the company may launch their first offline stores with the product that they may sell at their first offline store. Also, suggest the same for the online web store. Top cities on which they should focus the most and the type of apparel.
a.	Top categories and types
b.	Top products
c.	Top product type
2.	What insights can you offer about new and returning customers? 
a.	Analyze the old customers based on RFM(Recency, Frequency, and Monetary Value)
b.	Analyze the new customers.
c.	Find what the old and new customers are buying, at what time they are buying, and at what day of the month they are buying?
3.	What are some interesting insights that you can infer from the data?
4.	Is it fine to launch offers on weekdays or weekends?
5.	Which day of the month gets them the best sales?
6.	At what time of the day their website gets the maximum impressions? (Number of orders)
7.	Are men placing more orders than women?
8.	Is it a good idea for the company to open an offline store at this early stage of its startup journey?